requirements.txt

# MinkowskiEngine == 0.5.4
nuscenes-devkit == 1.1.9
# copy
tqdm
pytorch_lightning == 1.6.0
pyquaternion
# multiprocessing
scikit-image == 0.21.0
spconv == 2.1.21
scipy == 1.10.1
ftfy == 6.1.1
opencv-python == 4.5.1.48
setuptools == 59.5.0
imageio == 2.32.0
timm == 0.9.12
regex
einops
cython
shapely
h5py
submitit
wandb
numpy == 1.22.4
Pillow == 9.3.0
pycocotools ~= 2.0.4
fvcore
tabulate
opencv-python-headless
numba == 0.58.1
# pyyaml == 6.0.1
easydict == 1.11
open_clip_torch == 2.16.0